Move stories inside Storybook folder, run prettier over files
