+++ .storybook/main.js
@@ -1,9 +1,9 @@
 module.exports = {
-  stories: ['../stories/**/*.stories.js'],
+  stories: ['./stories/**/*.stories.js'],
   addons: [
     '@storybook/preset-typescript',
     '@storybook/addon-actions',
     '@storybook/addon-links',
-    '@storybook/addon-storysource'
+    '@storybook/addon-storysource',
   ],
 }
