+++ .storybook/theme.js
@@ -1,9 +1,9 @@
-import { create } from '@storybook/theming/create';
+import { create } from '@storybook/theming/create'
 
 import dreiLogo from './drei.jpeg'
 
 export default create({
-    base: "light",
-    brandImage: dreiLogo,
-    appBg: "white"
+  base: 'light',
+  brandImage: dreiLogo,
+  appBg: 'white',
 })
