+++ .storybook/useTurntable.js
@@ -1,13 +1,11 @@
-import React, { useRef } from 'react'
+import useRef from 'react'
 import { useFrame } from 'react-three-fiber'
 
 export function useTurntable() {
+  const ref = useRef()
+  useFrame(() => {
+    ref.current.rotation.y += 0.01
+  })
 
-    const ref = React.useRef()
-    useFrame(() => {
-        ref.current.rotation.y += 0.01
-    })
-
-    return ref
-
+  return ref
 }
